# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Gopher Social, a social media platform for Gophers.",
        "title": "Gopher Social API",
        "contact": {
            "name": "Rohit Vilas Ingole",
            "email": "rohit.vilas.ingole@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/DataRohit/Gopher-Social-Backend/blob/master/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/follow/{identifier}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows a logged-in user to follow another user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Follow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Identifier (username, email, or user ID) of the followee",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully followed user",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User account is inactive or banned",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Followee user not found",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Already following user",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to follow user",
                        "schema": {
                            "$ref": "#/definitions/models.FollowUserErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/followers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of users who are following the logged-in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Get followers of logged-in user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved followers list",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowersSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowersErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch followers",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowersErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/following": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of users that the logged-in user is following.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Get users being followed by logged-in user",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved following list",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowingSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowingErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch following users",
                        "schema": {
                            "$ref": "#/definitions/models.GetFollowingErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/unfollow/{identifier}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows a logged-in user to unfollow another user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Unfollow a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Identifier (username, email, or user ID) of the followee",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unfollowed user",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User account is inactive or banned",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Followee user not found or not following",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to unfollow user",
                        "schema": {
                            "$ref": "#/definitions/models.UnfollowUserErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{identifier}/followers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of users who are following the user identified by identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Get followers of a user by identifier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Identifier (username, email, or user ID) of the user",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved followers list for user",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowersSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowersErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowersErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not found",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowersErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch followers",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowersErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{identifier}/following": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of users that the user identified by identifier is following.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follow"
                ],
                "summary": "Get users being followed by a user by identifier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Identifier (username, email, or user ID) of the user",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved following list for user",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowingSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowingErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in or invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowingErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - User not found",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowingErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch following users",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserFollowingErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.FollowUserErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.FollowUserSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User Followed Successfully"
                }
            }
        },
...
        "models.GetFollowersErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.GetFollowersSuccessResponse": {
            "type": "object",
            "properties": {
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Followers Retrieved Successfully"
                }
            }
        },
        "models.GetFollowingErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.GetFollowingSuccessResponse": {
            "type": "object",
            "properties": {
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Following Users Retrieved Successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}